+++ examples/demo-rollup/provers/sp1/guest-celestia/src/main.rs
@@ -1,7 +1,6 @@
 #![no_main]
 sp1_zkvm::entrypoint!(main);
 
-use const_rollup_config::{ROLLUP_BATCH_NAMESPACE_RAW, ROLLUP_PROOF_NAMESPACE_RAW};
 use demo_stf::runtime::Runtime;
 use demo_stf::StfVerifier;
 use sov_address::MultiAddressEvm;
@@ -10,15 +9,15 @@ use sov_celestia_adapter::verifier::{CelestiaSpec, CelestiaVerifier};
 use sov_mock_zkvm::MockZkvm;
 use sov_modules_api::configurable_spec::ConfigurableSpec;
 use sov_modules_api::execution_mode::Zk;
+use sov_modules_api::macros::config_value;
 use sov_modules_stf_blueprint::StfBlueprint;
 use sov_rollup_interface::da::DaVerifier;
 use sov_sp1_adapter::guest::SP1Guest;
 use sov_sp1_adapter::SP1;
 use sov_state::ZkStorage;
 
-// The rollup stores its data in the namespace b"sov-test" on Celestia
-const ROLLUP_BATCH_NAMESPACE: Namespace = Namespace::const_v0(ROLLUP_BATCH_NAMESPACE_RAW);
-const ROLLUP_PROOF_NAMESPACE: Namespace = Namespace::const_v0(ROLLUP_PROOF_NAMESPACE_RAW);
+pub const ROLLUP_BATCH_NAMESPACE: Namespace = Namespace::const_v0(config_value!("BATCH_NAMESPACE"));
+pub const ROLLUP_PROOF_NAMESPACE: Namespace = Namespace::const_v0(config_value!("PROOF_NAMESPACE"));
 
 pub fn main() {
     let guest = SP1Guest::new();